Move `impl Controller<UnixStream>` to platform module
Move implementation of Controller for unix streams to platform
specific module.
