+++ Cargo.toml
@@ -10,6 +10,5 @@ regex = "0.1"
 rustc-serialize = "0.3"
 rust-crypto = "^0.2"
 rand = "0.3"
-unix_socket = "0.5.0"
 timer = "0.1.6"
 chrono = "0.2.22"
