+++ src/lib.rs
@@ -1,7 +1,6 @@
 #[macro_use]
 extern crate log;
 extern crate env_logger;
-extern crate unix_socket;
 extern crate regex;
 extern crate rustc_serialize;
 extern crate crypto;
@@ -9,6 +8,7 @@ extern crate rand;
 extern crate timer;
 extern crate chrono;
 
+pub mod platform;
 pub mod control;
 pub mod process;
 pub mod utils;
